Read method (Class App) done and created  new boolean variable checked in Message .class,, some methods on MessagePair class
Now read method is available with a new system of checking messages, when a user read a message the status of that messages changes from unchecked to checked, so the next time that the user read a message the previous one doenst appear.

Added a getter and setter to the new
boolean  variable checked and a getter for the user who send the message (from variable)

Added to message pair 2 methods to orgnise the read of the received messages and then to check them.