Update of method read()
Minor changes in the read() method and some refactoring done.

Also changed the way to update the status of the message: so now it seem to be unnecessary the method updateReceivedMessages(selectedIndex) because the message status changes everywhere (in all lists) directly.

Finally, in minor changes: some unnecessary variables or duplicated objects removed.