loadFromFile() and storeInFile() methods created in PrivateUser.java
These two methods are used to store a user data in a text file (userData.txt) in the project folder so the created user can be recovered from the file every time the app is launched.

This way when the new user is created the user's data is stored in the file.

To do: Create methods to store and load the server data...

Also "welcome" is printed only if there is an existing user... If not it will be created. So if the user is loaded correctly you will see "Welcome, yourNameHere".