New methods to Store and Load server data added.
new method seializeServer() added: it makes a byte code from the object Server, so it can be stored in a binary file server.ser. Later when user uses the app again, the server object will be recovered from the binary file (with the method deserializeServer() ) making it possible to access to all the data of the server's map added in previous sessions.

The binary file is updated every time any data of Server changes (after every send(), read(), createUser()...). Also the server is recovered every time the app is lounched.

NOTE:
Serialization is the process of converting an object's state into a byte stream so that it can be easily saved to a file, transmitted over a network, or stored in a database. This byte stream can later be used to recreate the original object.

Deserialization is the process of reconstructing an object from a byte stream. It involves reading the byte stream, interpreting the data, and recreating the original object with its state restored.